Move `KernelSwap` into `permutable_kernel`
